requirements.txt
numpy
scipy
matplotlib
pyyaml
numba
torch
